+++ src/sections/userpreset-browser.ts
@@ -51,9 +51,17 @@ export class UserPresetBrowser extends FavBrowserBase {
     // invoke base class method.
     super.render();
 
+    // filter items (if actions are not visible).
+    let filteredItems: Array<IUserPreset> | undefined;
+    if (!this.isActionsVisible) {
+      const filterName = (this.filterCriteria || "").toLocaleLowerCase();
+      filteredItems = this.mediaList?.filter((item: IUserPreset) => (item.name?.toLocaleLowerCase().indexOf(filterName) !== -1) || (item.subtitle?.toLocaleLowerCase().indexOf(filterName) !== -1));
+      this.filterItemCount = filteredItems?.length;
+    }
+
     // format title and sub-title details.
-    const title = formatTitleInfo(this.config.userPresetBrowserTitle, this.config, this.player, this.mediaListLastUpdatedOn, this.mediaList);
-    const subtitle = formatTitleInfo(this.config.userPresetBrowserSubTitle, this.config, this.player, this.mediaListLastUpdatedOn, this.mediaList);
+    const title = formatTitleInfo(this.config.userPresetBrowserTitle, this.config, this.player, this.mediaListLastUpdatedOn, this.mediaList, filteredItems);
+    const subtitle = formatTitleInfo(this.config.userPresetBrowserSubTitle, this.config, this.player, this.mediaListLastUpdatedOn, this.mediaList, filteredItems);
 
     // render html.
     return html`
@@ -70,12 +78,11 @@ export class UserPresetBrowser extends FavBrowserBase {
           ${(() => {
             // if actions are not visbile, then render the media list.
             if (!this.isActionsVisible) {
-              const filterName = (this.filterCriteria || "").toLocaleLowerCase();
               if (this.config.userPresetBrowserItemsPerRow === 1) {
                 return (
                   html`<spc-media-browser-list 
                         class="media-browser-list"
-                        .items=${this.mediaList?.filter((item: IUserPreset) => (item.name?.toLocaleLowerCase().indexOf(filterName) !== -1) || (item.subtitle?.toLocaleLowerCase().indexOf(filterName) !== -1))}
+                        .items=${filteredItems}
                         .store=${this.store}
                         @item-selected=${this.onItemSelected}
                         @item-selected-with-hold=${this.onItemSelectedWithHold}
@@ -85,7 +92,7 @@ export class UserPresetBrowser extends FavBrowserBase {
                 return (
                   html`<spc-media-browser-icons 
                         class="media-browser-list"
-                        .items=${this.mediaList?.filter((item: IUserPreset) => (item.name?.toLocaleLowerCase().indexOf(filterName) !== -1) || (item.subtitle?.toLocaleLowerCase().indexOf(filterName) !== -1))}
+                        .items=${filteredItems}
                         .store=${this.store}
                         @item-selected=${this.onItemSelected}
                         @item-selected-with-hold=${this.onItemSelectedWithHold}
@@ -401,12 +408,25 @@ export class UserPresetBrowser extends FavBrowserBase {
       // we use `finally` logic so we can clear the progress indicator.
       // any exceptions raised should have already been handled in the 
       // individual promise definitions; nothing else to do at this point.
-      Promise.allSettled(promiseRequests).finally(() => {
+      Promise.allSettled(promiseRequests)
+        .then(results => {
+          if (results) { }  // keep compiler happy
 
-        // clear the progress indicator.
-        this.progressHide();
+          //if (debuglog.enabled) {
+          //  debuglog("%cupdateMediaList - %s mediaList response:\n%s",
+          //    "color: red",
+          //    JSON.stringify(this.mediaType),
+          //    JSON.stringify(this.mediaList, null, 2),
+          //  );
+          //}
 
-      });
+        })
+        .finally(() => {
+
+          // clear the progress indicator.
+          this.progressHide();
+
+        });
 
       return true;